nba-prop-bot/requirements.txt
requests>=2.31.0
nba_api>=1.4.1
scipy>=1.11.0
numpy>=1.26.0
python-dotenv>=1.0.0
schedule>=1.2.0
